pyproject.toml
[project]
name = "senior-thesis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.10.1",
    "datasets>=2.14.0",
    "jax>=0.7.2",
    "numpy>=2.3.3",
    "peft>=0.7.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "tensorboard>=2.20.0",
    "torch>=2.5.1",
    "transformers>=4.30.0",
    "trl>=0.7.0",
]
